moved experiment, fixed experiment display
